requirements.txt
opencv-python
pillow
numpy
transformers
torch
qwen-vl-utils
